requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
boto3==1.26.79
botocore==1.29.82
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-silk==5.0.3
django-storages==1.13.2
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.9.0
frozenlist==1.3.3
gunicorn==20.1.0
httplib2==0.21.0
iniconfig==2.0.0
itypes==1.2.0
MarkupSafe==2.1.2
multidict==6.0.4
mypy==1.0.1
mypy-extensions==1.0.0
mysqlclient==2.1.1
nodeenv==1.7.0
packaging==23.0
Pillow==9.4.0
pre-commit==3.2.2
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
shortuuid==1.0.11
